It seems there is a problem with a filetered profile.
When the scorep plugin generated profile is loaded the Analyzer was to
seg fault. I added error handling in order to prevent it from crashing.
However, the general cause for this has not been solved. It seems as if
it has something to do with finding the main method in the filtered
profile.