Started working on the fix for a problem with polymorphic func calls
We discoveed that the problem we were knowing its there actually hits us
in DROPS. The problem is, that polymorphism is used at higher levels of
the call graph.
This leads to actually called nodes not being connected to the main
function. As a reslt those get pruned in the simplification estimation
phase.